Add files via upload
It is a small project,  simple and user-friendly To-Do list application using HTML, CSS, and JavaScript.

The main functionality of the application revolves around a single 'Add Task' button. If the user tries to add a task without entering any text, an alert message will be displayed, reminding them to enter a task description.

Once a task description is entered, it is added to the task queue in the list. Each task is accompanied by a cross symbol, which allows the user to cancel or delete a task if needed. To mark a task as completed, the user simply needs to click on the task, and it will be visually checked off as 'done'.

It also saves data even if the user closes the browser. This means that the user's task list will be preserved when they reopen the application.